Nuevas funciones y colores añadidos

+++ install.sh
@@ -12,20 +12,27 @@ echo "
 "
 #Variables para guardar si los programas están instalados o no.
 instalado_git=0
+instalado_dhcp=0
 instalado_mysql=0
 instalado_apache=0
 insladado_php=0
-instalado_dhcp=0
 instalado_dns=0
 instalado_nmap=0
 
+#Variables con la instalación de los programas.
+instalar_git="apt-get install git -y"
+instalar_dhcp="apt-get install isc-dhcp-server -y"
+instalar_dns="apt-get install bind9 -y"
+instalar_apache="apt-get install apache2 -y"
+instalar_php="apt-get install php libapache2-mod-php php-mysql -y"
+instalar_nmap="apt-get install nmap -y"
+instalar_mysql="apt-get install mysql-server -y"
 
 #Comprueba que está instalado el paquete de git y 
 #lo marca con color verde si está instalado y rojo si no lo está.
-echo -e "\e[32m"
-git --version
 echo -e "\e[0m"
-if [ $? -eq 0 ]; then
+#Comprueba que está instalado git.
+if [ -x "$(command -v git)" ]; then
     echo -e "\e[32mGit instalado\e[0m"
     instalado_git=1
 else
@@ -45,7 +52,7 @@ fi
 
 #Comprueba que está instalado los paquetes de PHP y PHP-MySQL
 #y lo marca con color verde si están instalados y rojo si no lo están.
-if [ -d /etc/php ]; then
+if [ -x "$(command -v php)" ]; then
     echo -e "\e[32mPHP instalado\e[0m"
     instalado_php=1
 else
@@ -97,7 +104,12 @@ fi
 suma_instalados=$(($instalado_git + $instalado_mysql + $instalado_apache + $instalado_php + $instalado_dhcp + $instalado_dns + $instalado_nmap))
 echo "****************************************************************************"
 if [ $suma_instalados -eq 7 ]; then
+    echo -e "\e[32m Todos los paquetes están instalados \e[0m"
+    echo -e "\e[35m Pulse cualquier tecla para continuar \e[0m"
+    # Al pulsar cualquier tecla se ejecuta el script de inicio.
+    read
     ./inicio.sh
+
 else
     #De color rojo.
     echo -e "\e[31m"
@@ -107,16 +119,18 @@ else
     if [ $opcion -eq 1 ]; then
         #Instala todos los paquetes.
         sudo apt-get update
+
         #Git
         if [ $instalado_git -eq 0 ]; then
-            echo "Instalando Git..."
-            sudo apt-get install git -y
+            #Color verde.
+            echo -e "\e[32m Instalando Git...\e[0m"
+            $instalar_git
         fi
-        #DHCP
 
+        #DHCP
         if [ $instalado_dhcp -eq 0 ]; then
-            echo "Instalando DHCP..."
-            sudo apt-get install isc-dhcp-server -y
+            echo -e "\e[32m Instalando DHCP...\e[0m"
+            $instalar_dhcp
             #Si se ha instalado cambia la variable a 1.
             if [ -f /etc/dhcp/dhcpd.conf ]; then
                 instalado_dhcp=1
@@ -126,8 +140,11 @@ else
         #Configurar DHCP (option domain-name "variable")
         #Ejemplo: option domain-name "instragram.com";
         if [ $instalado_dhcp -eq 1 ]; then
-            echo "Configurando DHCP..."
+            echo -e "\e[32m Configurando DHCP...\e[0m"
+            # En magenta.
+            echo -e "\e[35m"
             read -p "Introduzca el dominio: " dominio
+            echo -e "\e[0m"
             sed -i 's/option domain-name ".*";/option domain-name "'$dominio'";/g' /etc/dhcp/dhcpd.conf
         fi
 
@@ -136,7 +153,7 @@ else
         #search instragram.com
         #nameserver IP_Propia
         if [ $instalado_dhcp -eq 1 ]; then
-            echo "Configurando /etc/resolv.conf..."
+            echo -e "\e[32m Configurando /etc/resolv.conf...\e[0m"
             #EL DOMAIN NAME NO FUNCIONA AUN!!!!!!!!!!!!!!!!!!!!!!!!!
             sed -i 's/domain-name .*;/domain-name '$dominio';/g' /etc/resolv.conf
             sed -i 's/search.*/search '$dominio'/g' /etc/resolv.conf
@@ -145,8 +162,8 @@ else
 
         #DNS
         if [ $instalado_dns -eq 0 ]; then
-            echo "Instalando DNS..."
-            sudo apt-get install bind9 -y
+            echo -e "\e[32m Instalando DNS...\e[0m"
+            $instalar_dns
             #Si se ha instalado cambia la variable a 1.
             if [ -f /etc/bind/named.conf.options ]; then
                 instalado_dns=1
@@ -157,8 +174,11 @@ else
         #Copia el archivo /etc/bind/db.local a /etc/bind/db.$dominio
         #Ejemplo: db.$dominio
         if [ $instalado_dns -eq 1 ]; then
-            echo "Configurando DNS..."
+            echo -e "\e[32m Configurando DNS...\e[0m"
+            #En magenta.
+            echo -e "\e[35m"
             read -p "Introduzca el dominio para el DNS (el mismo de antes): " dominio
+            echo -e "\e[0m"
             cp /etc/bind/db.local /etc/bind/db.$dominio
             #Si db.$dominio existe modifica /etc/bind/db.$dominio.
             #IN SOA debianservidor.$dominio.com. root.debianservidor.$dominio.com. (
@@ -175,7 +195,7 @@ else
             #file "db.instragram.com";
             #};
             if [ -f /etc/bind/named.conf.local ]; then
-                echo "Configurando named.conf.local..."
+                echo -e "\e[32m Configurando named.conf.local...\e[0m"
                 echo "zone \"$dominio\" {" >> /etc/bind/named.conf.local
                 echo "type master;" >> /etc/bind/named.conf.local
                 echo "file \"db.$dominio\";" >> /etc/bind/named.conf.local
@@ -184,23 +204,84 @@ else
         fi
         #Apache2
         if [ $instalado_apache -eq 0 ]; then
-            echo "Instalando Apache2..."
-            apt-get install apache2 -y
+            echo -e "\e[32m Instalando Apache2...\e[0m"
+            $instalar_apache
         fi
         /etc/init.d/apache2 start
 
         #PHP
         if [ $instalado_php -eq 0 ]; then
-            echo "Instalando PHP..."
-            apt-get install php libapache2-mod-php php-mysql -y
+            echo -e "\e[32m Instalando PHP...\e[0m"
+            $instalar_php
         fi
         #Nmap
         if [ $instalado_nmap -eq 0 ]; then
-            echo "Instalando Nmap..."
-            apt-get install nmap -y
+            echo -e "\e[32m Instalando Nmap...\e[0m"
+            $instalar_nmap
         fi
+        #MySQL
+        if [ $instalado_mysql -eq 0 ]; then
+            echo -e "\e[32m Instalando MySQL...\e[0m"
+            $instalar_mysql
+        fi
+
+    #Elegir programa a instalar.
     elif [ $opcion -eq 2 ]; then
-        ./instalar.sh
+        echo "Seleccione una opcion: [1] Git [2] MySQL [3] Apache2 [4] PHP [5] DHCP [6] DNS [7] Nmap"
+        read -p "Opcion: " opcion
+        if [ $opcion -eq 1 ]; then
+            #Git
+            if [ $instalado_git -eq 0 ]; then
+                #Color verde.
+                echo -e "\e[32m Instalando Git...\e[0m"
+                $instalar_git
+            fi
+        elif [ $opcion -eq 2 ]; then
+            #MySQL
+            if [ $instalado_mysql -eq 0 ]; then
+                echo "Instalando MySQL..."
+                apt-get install mysql-server -y
+            fi
+        elif [ $opcion -eq 3 ]; then
+            #Apache2
+            if [ $instalado_apache -eq 0 ]; then
+                echo "Instalando Apache2..."
+                apt-get install apache2 -y
+            fi
+            /etc/init.d/apache2 start
+        elif [ $opcion -eq 4 ]; then
+            #PHP
+            if [ $instalado_php -eq 0 ]; then
+                echo "Instalando PHP..."
+                $instalar_php
+            fi
+        elif [ $opcion -eq 5 ]; then
+            #DHCP
+            if [ $instalado_dhcp -eq 0 ]; then
+                echo -e "\e[32m Instalando DHCP...\e[0m"
+                sudo apt-get install isc-dhcp-server -y
+                #Si se ha instalado cambia la variable a 1.
+                if [ -f /etc/dhcp/dhcpd.conf ]; then
+                    instalado_dhcp=1
+                fi
+            fi
+        elif [ $opcion -eq 6 ]; then
+            #DNS
+            if [ $instalado_dns -eq 0 ]; then
+                echo "Instalando DNS..."
+                sudo apt-get install
+                #Si se ha instalado cambia la variable a 1.
+                if [ -f /etc/bind/named.conf.options ]; then
+                    instalado_dns=1
+                fi
+            fi
+        elif [ $opcion -eq 7 ]; then
+            #Nmap
+            if [ $instalado_nmap -eq 0 ]; then
+                echo "Instalando Nmap..."
+                apt-get install nmap -y
+            fi
+        fi
     elif [ $opcion -eq 3 ]; then
         ./inicio.sh
     fi
